Move YKI-specific deserializers to separate package
